[FIX] 아이디 타입 안맞는것 수정

+++ backend/src/main/java/com/postrofit/backend/Model/DAO/StationDAO.java
@@ -14,7 +14,7 @@ import javax.persistence.*;
 public class StationDAO {
 
     @Id
-    private int stationId;
+    private long stationId;
 
     // 역 이름
     @Column
